Initial commit
While analyst should always login frequently, review, and categorize alerts, it may still be beneficial to be alerted when the Sguil uncategorized alert count becomes too high.  This script checks the number of uncategorized alerts in the Sguil database and sends an email alert when the number crosses the defined threshold.